Update 3.3 Domain Requirements
This app operates in the private educational field, and its main purpose is to facilitate the work of a
small company that courses in several foreign languages. The users who access the app have to be included in one of the login access categories. All users need to have a stable Internet
connection. This app manages all the operations related to our course center, but it also provides additional functions as well, such as calculating prices and getting notifications on special holidays, regarding grades and achievements. Different types of users have their own facilities. To log in as a student you also can give your own review so it can be shown to our app so we can help others decide about us. on the other hand, we have teachers who using our app will be able to connect virtually with their students. The manager using this app will be provided the opportunity to calculate and set the salaries of each employee, set prices, and check contracts.